refactor: replace literal env strings with Environment enum + restructure form component
- BE + UI specs: every test data fixture and toBe() assertion that used
  the literal strings 'unspecified' / 'production' / 'development' now
  reads from the Environment enum (imported from
  src/modules/database/entities/database.entity in BE and from apiClient
  in UI). Only line that stays a literal is the one inside a
  jest.mock(...) factory — jest's hoisting forbids referencing imports
  there.

- EnvironmentSelect rearranged per .ai/rules/frontend.md to one file per
  concern:
    form/EnvironmentSelect/
      EnvironmentSelect.tsx            (main component)
      EnvironmentSelect.types.ts       (Props)
      EnvironmentSelect.constants.ts   (ENVIRONMENT_OPTIONS)
      components/
        EnvironmentLabel.tsx           (label + tooltip trigger)
        EnvironmentTooltipContent.tsx  (tooltip body)
      index.ts                         (re-export default)

References: #RI-8198
