+++ redisinsight/api/src/modules/database/database.analytics.spec.ts
@@ -12,6 +12,7 @@ import { DEFAULT_SUMMARY as DEFAULT_REDIS_MODULES_SUMMARY } from 'src/utils/redi
 import { DatabaseAnalytics } from 'src/modules/database/database.analytics';
 import {
   Encoding,
+  Environment,
   HostingProvider,
 } from 'src/modules/database/entities/database.entity';
 
@@ -66,7 +67,7 @@ describe('DatabaseAnalytics', () => {
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           serverName: 'valkey',
           keyNameFormat: Encoding.UNICODE,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           ...DEFAULT_REDIS_MODULES_SUMMARY,
         },
       );
@@ -106,7 +107,7 @@ describe('DatabaseAnalytics', () => {
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           serverName: 'valkey',
           keyNameFormat: Encoding.UNICODE,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           ...DEFAULT_REDIS_MODULES_SUMMARY,
         },
       );
@@ -146,7 +147,7 @@ describe('DatabaseAnalytics', () => {
           forceStandalone: 'false',
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           keyNameFormat: Encoding.UNICODE,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           ...DEFAULT_REDIS_MODULES_SUMMARY,
           RediSearch: {
             loaded: true,
@@ -193,7 +194,7 @@ describe('DatabaseAnalytics', () => {
           forceStandalone: 'false',
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           keyNameFormat: Encoding.UNICODE,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           ...DEFAULT_REDIS_MODULES_SUMMARY,
           RediSearch: {
             loaded: true,
@@ -254,7 +255,7 @@ describe('DatabaseAnalytics', () => {
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           forceStandalone: 'false',
           keyNameFormat: Encoding.UNICODE,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           previousValues: {
             connectionType: prev.connectionType,
             provider: prev.provider,
@@ -265,7 +266,7 @@ describe('DatabaseAnalytics', () => {
             useSSH: 'disabled',
             forceStandalone: 'false',
             keyNameFormat: Encoding.UNICODE,
-            environment: 'unspecified',
+            environment: Environment.Unspecified,
           },
         },
       );
@@ -297,7 +298,7 @@ describe('DatabaseAnalytics', () => {
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           forceStandalone: 'false',
           keyNameFormat: Encoding.UNICODE,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           previousValues: {
             connectionType: prev.connectionType,
             provider: prev.provider,
@@ -308,7 +309,7 @@ describe('DatabaseAnalytics', () => {
             useTLSAuthClients: 'disabled',
             forceStandalone: 'false',
             keyNameFormat: Encoding.UNICODE,
-            environment: 'unspecified',
+            environment: Environment.Unspecified,
           },
         },
       );
@@ -340,7 +341,7 @@ describe('DatabaseAnalytics', () => {
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           forceStandalone: 'true',
           keyNameFormat: Encoding.UNICODE,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           previousValues: {
             connectionType: prev.connectionType,
             provider: prev.provider,
@@ -351,7 +352,7 @@ describe('DatabaseAnalytics', () => {
             useTLSAuthClients: 'disabled',
             forceStandalone: 'false',
             keyNameFormat: Encoding.UNICODE,
-            environment: 'unspecified',
+            environment: Environment.Unspecified,
           },
         },
       );
@@ -383,7 +384,7 @@ describe('DatabaseAnalytics', () => {
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           forceStandalone: 'false',
           keyNameFormat: Encoding.UNICODE,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           previousValues: {
             connectionType: prev.connectionType,
             provider: prev.provider,
@@ -394,7 +395,7 @@ describe('DatabaseAnalytics', () => {
             useTLSAuthClients: 'disabled',
             forceStandalone: 'false',
             keyNameFormat: Encoding.UNICODE,
-            environment: 'unspecified',
+            environment: Environment.Unspecified,
           },
         },
       );
@@ -427,7 +428,7 @@ describe('DatabaseAnalytics', () => {
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           forceStandalone: 'true',
           keyNameFormat: Encoding.UNICODE,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           previousValues: {
             connectionType: prev.connectionType,
             provider: prev.provider,
@@ -438,7 +439,7 @@ describe('DatabaseAnalytics', () => {
             useTLSAuthClients: 'disabled',
             forceStandalone: 'true',
             keyNameFormat: Encoding.UNICODE,
-            environment: 'unspecified',
+            environment: Environment.Unspecified,
           },
         },
       );
@@ -473,7 +474,7 @@ describe('DatabaseAnalytics', () => {
           useDecompression: mockDatabaseWithTlsAuth.compressor,
           forceStandalone: 'false',
           keyNameFormat: Encoding.HEX,
-          environment: 'unspecified',
+          environment: Environment.Unspecified,
           previousValues: {
             connectionType: prev.connectionType,
             provider: prev.provider,
@@ -484,7 +485,7 @@ describe('DatabaseAnalytics', () => {
             useSSH: 'disabled',
             forceStandalone: 'false',
             keyNameFormat: Encoding.UNICODE,
-            environment: 'unspecified',
+            environment: Environment.Unspecified,
           },
         },
       );
