+++ redisinsight/ui/src/pages/home/components/form/EnvironmentSelect/EnvironmentSelect.constants.ts
@@ -1,0 +1,7 @@
+import { Environment } from 'apiClient'
+
+export const ENVIRONMENT_OPTIONS = [
+  { value: Environment.Unspecified, label: 'Unspecified' },
+  { value: Environment.Production, label: 'Production' },
+  { value: Environment.Development, label: 'Development' },
+]
