+++ redisinsight/ui/src/pages/home/components/form/EnvironmentSelect/EnvironmentSelect.types.ts
@@ -1,0 +1,7 @@
+import { FormikProps } from 'formik'
+
+import { DbConnectionInfo } from 'uiSrc/pages/home/interfaces'
+
+export interface EnvironmentSelectProps {
+  formik: FormikProps<DbConnectionInfo>
+}
