+++ redisinsight/ui/src/pages/home/components/form/EnvironmentSelect/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EnvironmentSelect'
